Backend/requirements.txt
Flask
psycopg2-binary
Flask-CORS
SQLAlchemy
python-dotenv
flask-migrate
flask_sqlalchemy
Flask-Session
redis